+++ snapshots/output/basic-project/lib/more.dart
@@ -2,15 +2,11 @@
 // definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/
 //                      ^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/math.
   
-  enum AnimalType {
+  enum AnimalType { cat, dog, bird }
 //     ^^^^^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#
-    cat,
-//  ^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#cat.
-    dog,
-//  ^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#dog.
-    bird,
-//  ^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#bird.
-  }
+//                  ^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#cat.
+//                       ^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#dog.
+//                            ^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#bird.
   
   typedef SoundMaker = void Function();
 //        ^^^^^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/SoundMaker#
@@ -20,7 +16,7 @@
     void sleep() {
 //       ^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/SleepMixin#sleep().
       print('zzz...');
-//    ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//    ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
     }
   }
   
@@ -29,7 +25,7 @@
 //      relationship scip-dart pub dart_test 1.0.0 lib/`more.dart`/SleepMixin# implementation
 //                  ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/SleepMixin#
     String name;
-//  ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`string.dart`/String#
+//  ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`string.dart`/String#
 //         ^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#name.
     AnimalType type;
 //  ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#
@@ -44,7 +40,7 @@
 //                                   ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#type.
 //                                   ^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#`<constructor>`().(type)
       print(AnimalType.values);
-//    ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//    ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
 //          ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#
       switch (type) {
 //            ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#type.
@@ -53,26 +49,26 @@
 //                      ^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#cat.
           soundMaker = () => print('Meow!');
 //        ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#soundMaker.
-//                           ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//                           ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
           break;
         case AnimalType.dog:
 //           ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#
 //                      ^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#dog.
           soundMaker = () => print('Woof!');
 //        ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#soundMaker.
-//                           ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//                           ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
           break;
         case AnimalType.bird:
 //           ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#
 //                      ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/AnimalType#bird.
           soundMaker = () => print('Chirp!');
 //        ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#soundMaker.
-//                           ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//                           ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
           break;
         default:
           soundMaker = () => print('Unknown animal type');
 //        ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#soundMaker.
-//                           ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//                           ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
       }
     }
   
@@ -90,11 +86,11 @@
     }
   
     @override
-//   ^^^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`annotations.dart`/override.
+//   ^^^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`annotations.dart`/override.
     String toString() {
-//  ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`string.dart`/String#
+//  ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`string.dart`/String#
 //         ^^^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#toString().
-//         relationship scip-dart pub dart:core 2.19.0 dart:core/`object.dart`/Object#toString(). implementation reference
+//         relationship scip-dart pub dart:core 3.13.0 dart:core/`object.dart`/Object#toString(). implementation reference
       return '$name the $type';
 //             ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#name.
 //                       ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#type.
@@ -102,14 +98,14 @@
   }
   
   int calculateSum(List<int> numbers) {
-//^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`int.dart`/int#
+//^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`int.dart`/int#
 //    ^^^^^^^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/calculateSum().
-//                 ^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`list.dart`/List#
-//                      ^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`int.dart`/int#
+//                 ^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`list.dart`/List#
+//                      ^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`int.dart`/int#
 //                           ^^^^^^^ definition local 0
     return numbers.reduce((value, element) => value + element);
 //         ^^^^^^^ reference local 0
-//                 ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`iterable.dart`/Iterable#reduce().
+//                 ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`iterable.dart`/Iterable#reduce().
 //                         ^^^^^ definition local 1
 //                                ^^^^^^^ definition local 2
 //                                            ^^^^^ reference local 1
@@ -119,11 +115,11 @@
   void main() {
 //     ^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/main().
     List<int> numbers = [1, 2, 3, 4, 5];
-//  ^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`list.dart`/List#
-//       ^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`int.dart`/int#
+//  ^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`list.dart`/List#
+//       ^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`int.dart`/int#
 //            ^^^^^^^ definition local 3
     int sum = calculateSum(numbers);
-//  ^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`int.dart`/int#
+//  ^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`int.dart`/int#
 //      ^^^ definition local 4
 //            ^^^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/calculateSum().
 //                         ^^^^^^^ reference local 3
@@ -167,36 +163,72 @@
 //      ^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/Animal#makeSound().
   
     print(bird);
-//  ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//  ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
 //        ^^^^ reference local 5
     print(dog);
-//  ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//  ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
 //        ^^^ reference local 6
     print('The sum of $numbers is $sum');
-//  ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//  ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
 //                     ^^^^^^^ reference local 3
 //                                 ^^^ reference local 4
   
-    print(math.Rectangle(1,2,3,4));
-//  ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+    print(math.Rectangle(1, 2, 3, 4));
+//  ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
 //        ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/math.
-//             ^^^^^^^^^ reference scip-dart pub dart:math 2.19.0 dart:math/`rectangle.dart`/Rectangle#`<constructor>`().
+//             ^^^^^^^^^ reference scip-dart pub dart:math 3.13.0 dart:math/`rectangle.dart`/Rectangle#`<constructor>`().
   
-    [1,2].reduce((a, b) => a + b);
-//        ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`iterable.dart`/Iterable#reduce().
-//                ^ definition local 8
-//                   ^ definition local 9
-//                         ^ reference local 8
-//                             ^ reference local 9
+    math.Random? random = math.Random();
+//  ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/math.
+//       ^^^^^^ reference scip-dart pub dart:math 3.13.0 dart:math/`random.dart`/Random#
+//               ^^^^^^ definition local 8
+//                        ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/math.
+//                             ^^^^^^ reference scip-dart pub dart:math 3.13.0 dart:math/`random.dart`/Random#`<constructor>`().
+    final points = <math.Point<int>>[];
+//        ^^^^^^ definition local 9
+//                  ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/math.
+//                       ^^^^^ reference scip-dart pub dart:math 3.13.0 dart:math/`point.dart`/Point#
+//                             ^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`int.dart`/int#
+    Object value = points;
+//  ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`object.dart`/Object#
+//         ^^^^^ definition local 10
+//                 ^^^^^^ reference local 9
+    print(value is math.Random);
+//  ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
+//        ^^^^^ reference local 10
+//                 ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/math.
+//                      ^^^^^^ reference scip-dart pub dart:math 3.13.0 dart:math/`random.dart`/Random#
+    print([random, points]);
+//  ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
+//         ^^^^^^ reference local 8
+//                 ^^^^^^ reference local 9
+  
+    [1, 2].reduce((a, b) => a + b);
+//         ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`iterable.dart`/Iterable#reduce().
+//                 ^ definition local 11
+//                    ^ definition local 12
+//                          ^ reference local 11
+//                              ^ reference local 12
   }
   
   void test(String Function(int) p) {}
 //     ^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/test().
-//          ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`string.dart`/String#
-//                          ^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`int.dart`/int#
-//                               ^ definition local 10
+//          ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`string.dart`/String#
+//                          ^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`int.dart`/int#
+//                               ^ definition local 13
   void deepTest(String Function(void Function(String test)) p) {}
 //     ^^^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/deepTest().
-//              ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`string.dart`/String#
-//                                            ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`string.dart`/String#
-//                                                          ^ definition local 11
+//              ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`string.dart`/String#
+//                                            ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`string.dart`/String#
+//                                                          ^ definition local 14
+  math.Point<int> origin(math.Random random) => math.Point(0, 0);
+//^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/math.
+//     ^^^^^ reference scip-dart pub dart:math 3.13.0 dart:math/`point.dart`/Point#
+//           ^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`int.dart`/int#
+//                ^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`more.dart`/origin().
+//                       ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/math.
+//                            ^^^^^^ reference scip-dart pub dart:math 3.13.0 dart:math/`random.dart`/Random#
+//                                   ^^^^^^ definition local 15
+//                                              ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/math.
+//                                                   ^^^^^ reference scip-dart pub dart:math 3.13.0 dart:math/`point.dart`/Point#`<constructor>`().
+