+++ snapshots/output/basic-project/lib/other.dart
@@ -5,33 +5,28 @@
   class Foo {
 //      ^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#
     int _far;
-//  ^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`int.dart`/int#
+//  ^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`int.dart`/int#
 //      ^^^^ definition local 0
     bool value;
-//  ^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`bool.dart`/bool#
+//  ^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`bool.dart`/bool#
 //       ^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#value.
     String value2;
-//  ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`string.dart`/String#
+//  ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`string.dart`/String#
 //         ^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#value2.
     double? value3;
-//  ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`double.dart`/double#
+//  ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`double.dart`/double#
 //          ^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#value3.
-    Foo(
+    Foo(this._far, {required this.value, required this.value2, this.value3}) {
 //  ^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#`<constructor>`().
-      this._far, {
-//         ^^^^ reference local 0
-      required this.value,
-//                  ^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#value.
-//                  ^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#`<constructor>`().(value)
-      required this.value2,
-//                  ^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#value2.
-//                  ^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#`<constructor>`().(value2)
-      this.value3,
-//         ^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#value3.
-//         ^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#`<constructor>`().(value3)
-    }) {
+//           ^^^^ reference local 0
+//                                ^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#value.
+//                                ^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#`<constructor>`().(value)
+//                                                     ^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#value2.
+//                                                     ^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#`<constructor>`().(value2)
+//                                                                  ^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#value3.
+//                                                                  ^^^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#`<constructor>`().(value3)
       print(_far);
-//    ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//    ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
 //          ^^^^ reference local 0
     }
   }
@@ -39,7 +34,7 @@
   class Bar {
 //      ^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Bar#
     String _someValue;
-//  ^^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`string.dart`/String#
+//  ^^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`string.dart`/String#
 //         ^^^^^^^^^^ definition local 1
     Bar(this._someValue);
 //  ^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/Bar#`<constructor>`().
@@ -50,21 +45,19 @@
       _someValue = 'asdf';
 //    ^^^^^^^^^^ reference local 1
       print(_someValue);
-//    ^^^^^ reference scip-dart pub dart:core 2.19.0 dart:core/`print.dart`/print().
+//    ^^^^^ reference scip-dart pub dart:core 3.13.0 dart:core/`print.dart`/print().
 //          ^^^^^^^^^^ reference local 1
     }
   }
   
   void main() {
 //     ^^^^ definition scip-dart pub dart_test 1.0.0 lib/`other.dart`/main().
-    more.loadLibrary().then((_) => {
+    more.loadLibrary().then((_) => {Bar('a').someMethod.call()});
 //  ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/more.
-//                     ^^^^ reference scip-dart pub dart:async 2.19.0 dart:async/`future.dart`/Future#then().
+//                     ^^^^ reference scip-dart pub dart:async 3.13.0 dart:async/`future.dart`/Future#then().
 //                           ^ definition local 2
-      Bar('a').someMethod.call()
-//    ^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Bar#`<constructor>`().
-//             ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Bar#someMethod().
-    });
+//                                  ^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Bar#`<constructor>`().
+//                                           ^^^^^^^^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Bar#someMethod().
   
     Foo(1, value: true, value2: 'asdf')..value = false;
 //  ^^^ reference scip-dart pub dart_test 1.0.0 lib/`other.dart`/Foo#`<constructor>`().
@@ -91,3 +84,4 @@
 //       ^^^^ reference scip-dart pub dart_test 1.0.0 lib/`more.dart`/test().
 //             ^ definition local 4
   }
+